+++ iCarousel/iCarousel.m
@@ -1,7 +1,7 @@
 //
 //  iCarousel.m
 //
-//  Version 1.7.4
+//  Version 1.7.5
 //
 //  Created by Nick Lockwood on 01/04/2011.
 //  Copyright 2011 Charcoal Design
@@ -48,11 +48,13 @@
 #if __has_feature(objc_arc)
 #define ah_retain self
 #define ah_dealloc self
-#define release self
-#define autorelease self
+#define ah_release self
+#define ah_autorelease self
 #else
 #define ah_retain retain
 #define ah_dealloc dealloc
+#define ah_release release
+#define ah_autorelease autorelease
 #define __bridge
 #endif
 #endif
@@ -110,50 +112,6 @@ NSComparisonResult compareViewDepth(UIView *view1, UIView *view2, iCarousel *sel
 
 @implementation iCarousel
 
-@synthesize dataSource = _dataSource;
-@synthesize delegate = _delegate;
-@synthesize type = _type;
-@synthesize perspective = _perspective;
-@synthesize numberOfItems = _numberOfItems;
-@synthesize numberOfPlaceholders = _numberOfPlaceholders;
-@synthesize numberOfPlaceholdersToShow = _numberOfPlaceholdersToShow;
-@synthesize numberOfVisibleItems = _numberOfVisibleItems;
-@synthesize contentView = _contentView;
-@synthesize itemViews = _itemViews;
-@synthesize itemViewPool = _itemViewPool;
-@synthesize placeholderViewPool = _placeholderViewPool;
-@synthesize previousItemIndex = _previousItemIndex;
-@synthesize itemWidth = _itemWidth;
-@synthesize scrollOffset = _scrollOffset;
-@synthesize offsetMultiplier = _offsetMultiplier;
-@synthesize startVelocity = _startVelocity;
-@synthesize timer = _timer;
-@synthesize decelerating = _decelerating;
-@synthesize scrollEnabled = _scrollEnabled;
-@synthesize decelerationRate = _decelerationRate;
-@synthesize bounceDistance = _bounceDistance;
-@synthesize bounces = _bounces;
-@synthesize contentOffset = _contentOffset;
-@synthesize viewpointOffset = _viewpointOffset;
-@synthesize startOffset = _startOffset;
-@synthesize endOffset = _endOffset;
-@synthesize scrollDuration = _scrollDuration;
-@synthesize startTime = _startTime;
-@synthesize scrolling = _scrolling;
-@synthesize previousTranslation = _previousTranslation;
-@synthesize wrapEnabled = _wrapEnabled;
-@synthesize vertical = _vertical;
-@synthesize dragging = _dragging;
-@synthesize didDrag = _didDrag;
-@synthesize scrollSpeed = _scrollSpeed;
-@synthesize toggleTime = _toggleTime;
-@synthesize toggle = _toggle;
-@synthesize stopAtItemBoundary = _stopAtItemBoundary;
-@synthesize scrollToItemBoundary = _scrollToItemBoundary;
-@synthesize ignorePerpendicularSwipes = _ignorePerpendicularSwipes;
-@synthesize animationDisableCount = _animationDisableCount;
-@synthesize centerItemWhenSelected = _centerItemWhenSelected;
-
 #pragma mark -
 #pragma mark Initialisation
 
@@ -184,7 +142,7 @@ NSComparisonResult compareViewDepth(UIView *view1, UIView *view2, iCarousel *sel
     UIPanGestureRecognizer *panGesture = [[UIPanGestureRecognizer alloc] initWithTarget:self action:@selector(didPan:)];
     panGesture.delegate = (id <UIGestureRecognizerDelegate>)self;
     [_contentView addGestureRecognizer:panGesture];
-    [panGesture release];
+    [panGesture ah_release];
     
 #else
     
@@ -227,10 +185,10 @@ NSComparisonResult compareViewDepth(UIView *view1, UIView *view2, iCarousel *sel
 {   
     [self stopAnimation];
     
-    [_contentView release];
-    [_itemViews release];
-    [_itemViewPool release];
-    [_placeholderViewPool release];
+    [_contentView ah_release];
+    [_itemViews ah_release];
+    [_itemViewPool ah_release];
+    [_placeholderViewPool ah_release];
     [super ah_dealloc];
 }
 
@@ -866,7 +824,7 @@ NSComparisonResult compareViewDepth(UIView *view1, UIView *view2, iCarousel *sel
     CGRect frame = view.bounds;
     frame.size.width = _vertical? frame.size.width: _itemWidth;
     frame.size.height = _vertical? _itemWidth: frame.size.height;
-    UIView *containerView = [[[UIView alloc] initWithFrame:frame] autorelease];
+    UIView *containerView = [[[UIView alloc] initWithFrame:frame] ah_autorelease];
     
 #ifdef ICAROUSEL_IOS
     
@@ -874,7 +832,7 @@ NSComparisonResult compareViewDepth(UIView *view1, UIView *view2, iCarousel *sel
     UITapGestureRecognizer *tapGesture = [[UITapGestureRecognizer alloc] initWithTarget:self action:@selector(didTap:)];
     tapGesture.delegate = (id <UIGestureRecognizerDelegate>)self;
     [containerView addGestureRecognizer:tapGesture];
-    [tapGesture release];
+    [tapGesture ah_release];
     
 #else
     
@@ -1225,7 +1183,7 @@ NSComparisonResult compareViewDepth(UIView *view1, UIView *view2, iCarousel *sel
     {
         [_itemViewPool removeObject:view];
     }
-    return [view autorelease];
+    return [view ah_autorelease];
 }
 
 - (UIView *)dequeuePlaceholderView
@@ -1235,7 +1193,7 @@ NSComparisonResult compareViewDepth(UIView *view1, UIView *view2, iCarousel *sel
     {
         [_placeholderViewPool removeObject:view];
     }
-    return [view autorelease];
+    return [view ah_autorelease];
 }
 
 
@@ -1262,7 +1220,7 @@ NSComparisonResult compareViewDepth(UIView *view1, UIView *view2, iCarousel *sel
     
     if (view == nil)
     {
-        view = [[[UIView alloc] init] autorelease];
+        view = [[[UIView alloc] init] ah_autorelease];
     }
     [self setItemView:view forIndex:index];
     if (containerView)
@@ -1574,6 +1532,7 @@ NSComparisonResult compareViewDepth(UIView *view1, UIView *view2, iCarousel *sel
         [UIView commitAnimations];
         
         [UIView beginAnimations:nil context:nil];
+        [UIView setAnimationDelay:0.1];
         [UIView setAnimationDuration:INSERT_DURATION];
         [UIView setAnimationDelegate:self];
         [UIView setAnimationDidStopSelector:@selector(depthSortViews)];
@@ -1734,11 +1693,19 @@ NSComparisonResult compareViewDepth(UIView *view1, UIView *view2, iCarousel *sel
 {
     if (!_timer)
     {
-        _timer = [NSTimer scheduledTimerWithTimeInterval:1.0/60.0
-                                                  target:self
-                                                selector:@selector(step)
-                                                userInfo:nil
-                                                 repeats:YES];
+        self.timer = [NSTimer timerWithTimeInterval:1.0/60.0
+                                             target:self
+                                           selector:@selector(step)
+                                           userInfo:nil
+                                            repeats:YES];
+        
+        [[NSRunLoop mainRunLoop] addTimer:_timer forMode:NSDefaultRunLoopMode];
+
+#ifdef ICAROUSEL_IOS
+        
+        [[NSRunLoop mainRunLoop] addTimer:_timer forMode:UITrackingRunLoopMode];
+#endif
+        
     }
 }